pkg/transformer/eth_getLogs.go: move extractETHLogsFromTransactionReceipt() to util.go
