+++ pkg/transformer/eth_getLogs.go
@@ -3,7 +3,6 @@ package transformer
 import (
 	"encoding/json"
 
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/qtumproject/janus/pkg/eth"
 	"github.com/qtumproject/janus/pkg/qtum"
 	"github.com/qtumproject/janus/pkg/utils"
@@ -87,25 +86,4 @@ func (p *ProxyETHGetLogs) ToRequest(ethreq *eth.GetLogsRequest) (*qtum.SearchLog
 		FromBlock: from,
 		ToBlock:   to,
 	}, nil
-}
-
-func extractETHLogsFromTransactionReceipt(receipt *qtum.TransactionReceipt) []eth.Log {
-	logs := make([]eth.Log, 0, len(receipt.Log))
-	for i, log := range receipt.Log {
-		topics := make([]string, 0, len(log.Topics))
-		for _, topic := range log.Topics {
-			topics = append(topics, utils.AddHexPrefix(topic))
-		}
-		logs = append(logs, eth.Log{
-			TransactionHash:  utils.AddHexPrefix(receipt.TransactionHash),
-			TransactionIndex: hexutil.EncodeUint64(receipt.TransactionIndex),
-			BlockHash:        utils.AddHexPrefix(receipt.BlockHash),
-			BlockNumber:      hexutil.EncodeUint64(receipt.BlockNumber),
-			Data:             utils.AddHexPrefix(log.Data),
-			Address:          utils.AddHexPrefix(log.Address),
-			Topics:           topics,
-			LogIndex:         hexutil.EncodeUint64(uint64(i)),
-		})
-	}
-	return logs
 }
